requirements.txt
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
awsebcli==3.20.3
backcall==0.2.0
bcrypt==3.2.2
black==22.3.0
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.3
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
dj-database-url==0.5.0
Django==3.2.14
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
executing==0.8.3
future==0.16.0
gunicorn==20.1.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==7.2.2
jupyter-core==4.10.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.0
packaging==21.3
paramiko==2.11.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
texttable==1.6.4
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.10
wcwidth==0.1.9
websocket-client==0.59.0
whitenoise==6.2.0
